TIP-1287: move product model queries
