+++ src/Akeneo/Pim/Enrichment/Bundle/Storage/Sql/ProductModel/CountProductModelsAndChildrenProductModels.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Enrichment\Bundle\ProductModel\Query\Sql;
+namespace Akeneo\Pim\Enrichment\Bundle\Storage\Sql\ProductModel;
 
 use Akeneo\Pim\Enrichment\Component\Product\ProductModel\Query\CountProductModelsAndChildrenProductModelsInterface;
 use Doctrine\DBAL\Connection;
