+++ src/Akeneo/Pim/Enrichment/Bundle/Storage/Sql/ProductModel/GetProductAssociationsByProductModelCodes.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace Akeneo\Pim\Enrichment\Bundle\ProductModel\Query\Sql;
+namespace Akeneo\Pim\Enrichment\Bundle\Storage\Sql\ProductModel;
 
 use Doctrine\DBAL\Connection;
